requirements.txt
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
